[Improve][Zeta] Optimized code
