+++ config/seatunnel.yaml
@@ -39,7 +39,6 @@ seatunnel:
         enabled: false
       log:
         scheduled-deletion-enable: false
-        prefix: job
     http:
       enable-http: true
       port: 8080
