+++ seatunnel-e2e/seatunnel-engine-e2e/connector-seatunnel-e2e-base/src/test/resources/cluster/seatunnel.yaml
@@ -40,4 +40,3 @@ seatunnel:
              enabled: false
           logs:
              scheduled-deletion-enable: true
-             prefix: job
