+++ seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/SeaTunnelServer.java
@@ -146,8 +146,7 @@ public class SeaTunnelServer
                 && seaTunnelConfig.getEngineConfig().getTelemetryConfig().getLogs().isEnabled()) {
             taskLogManagerService =
                     new TaskLogManagerService(
-                            seaTunnelConfig.getEngineConfig().getTelemetryConfig().getLogs(),
-                            nodeEngine);
+                            seaTunnelConfig.getEngineConfig().getTelemetryConfig().getLogs());
             taskLogManagerService.initClean();
         }
